% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-abtestFunctions.R
\name{ab_test}
\alias{ab_test}
\title{Function for conducting a Bayesian A/B test.}
\usage{
ab_test(data, prior_par, nsamples = 10000, is_df = 5)
}
\arguments{
\item{data}{data}

\item{prior_par}{list with prior parameters}

\item{nsamples}{number of samples for importance sampling}

\item{is_df}{importance sampling degrees of freedom}
}
\description{
Function for conducting a Bayesian A/B test.
}
