% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{BFDA.analysis}
\alias{BFDA.analysis}
\title{Analyze aBFDA.sim object}
\usage{
BFDA.analysis(BFDA, n.min = NA, n.max = NA, boundary = NA,
  verbose = TRUE, alpha = 0.05)
}
\arguments{
\item{BFDA}{The result object from a BFDA.sim function}

\item{n.min}{What is the minimum n that is sampled before optional stopping is started? Defaults to the smallest n in the BFDA object}

\item{n.max}{What is the minimum n that is sampled before optional stopping is started? Defaults to the largest n in the BFDA object.}

\item{boundary}{At which BF boundary should trajectories stop? Either a single number (then the reciprocal is taken as the other boundary), or a vector of two numbers for lower and upper boundary.}

\item{verbose}{Print information about analysis?}

\item{alpha}{For a frequentist analysis in the fixed-n case: Use this alpha level.}
}
\description{
Analyze aBFDA.sim object
}
\examples{
\dontrun{
BFDA.analysis(sim, boundary=6, n.max=80)
}
}

