% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{BFDA.analyze}
\alias{BFDA.analyze}
\title{Analyze a BFDA.sim object}
\usage{
BFDA.analyze(BFDA, design = c("sequential", "fixed"), boundary,
  n.min = NULL, n.max = NULL, n = NULL, verbose = TRUE,
  alpha = 0.05)
}
\arguments{
\item{BFDA}{The result object from a \code{\link{BFDA.sim}} function call}

\item{design}{The function can analyze the simulations with regard to a fixed-n (\code{design="fixed"}) or in a sequential style (\code{design="sequential"})}

\item{boundary}{At which BF boundary should sequential trajectories stop? Either a single number (then the reciprocal is taken as the other boundary), or a vector of two numbers for lower and upper boundary (only relevant for \code{design="sequential"}).}

\item{n.min}{What is the minimum n that is sampled before optional stopping is started? Defaults to the smallest n in the BFDA object (only relevant for \code{design="sequential"}).}

\item{n.max}{What is the maximum n that is sampled in sequential sampling? Defaults to the largest n in the BFDA object (only relevant for \code{design="sequential"}).}

\item{n}{The fixed-n sample size that should be evaluated (only relevant for \code{design="fixed"})}

\item{verbose}{Print information about analysis?}

\item{alpha}{For a frequentist analysis in the fixed-n case: Use this alpha level.}
}
\description{
Analyze a BFDA.sim object
}
\examples{
\dontrun{
BFDA.analyze(sim, design="sequential", boundary=6, n.min=20, n.max=80)
}
}
