% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{BPA.analysis}
\alias{BPA.analysis}
\title{Analyze aBPA.sim object}
\usage{
BPA.analysis(BPA, n.max = NA, boundary = NA, verbose = TRUE,
  alpha = 0.05)
}
\arguments{
\item{BPA}{The result object from a BPA.sim function}

\item{n.max}{At which maximum n should the simulations be analyzed?}

\item{boundary}{At which BF boundary should trajectories stop?}

\item{verbose}{Print information about analysis?}

\item{alpha}{For a frequentist analysis in the fixed-n case: Use this alpha level.}
}
\description{
Analyze aBPA.sim object
}
\examples{
\dontrun{
BPA.analysis(sim, boundary=6, n.max=80)
}
}

