% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtest.R
\name{freq.test.abtest}
\alias{freq.test.abtest}
\title{freq.test.function: return p.value, test statistic, and empirical ES}
\usage{
freq.test.abtest(SAMP, alternative = NULL, options.sample = NULL)
}
\arguments{
\item{SAMP}{(reduced) sample}

\item{alternative}{direction of the alternative hypothesis}

\item{options.sample}{list(effecttype = ...)}
}
\description{
freq.test.function: return p.value, test statistic, and empirical ES
}
