% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-Plot.R
\name{plotBPA}
\alias{plot.BPA}
\alias{plotBPA}
\title{Plot a BPA object}
\usage{
plotBPA(BPA, boundary = 30, n.trajectories = 60, n.max = NA,
  dens.amplification = 1, cat = 3, bw = FALSE, dens.right.offset = 2,
  xlim = NA, ylim = NA, xextension = 1.5,
  traj.selection = "proportional", yaxis.at = NA, yaxis.labels = NA,
  forH1 = TRUE, n.max.label.position = "dynamic")
}
\arguments{
\item{BPA}{A BPA simulation object, resulting from the \code{BPA.sim} function}

\item{boundary}{The BF threshold where to stop (resp. its reciprocal)}

\item{n.trajectories}{Number of demo trajectories}

\item{n.max}{Maximum sample size}

\item{dens.amplification}{Amplification factor for all densities.}

\item{cat}{how many BF categories should be colored in the right density? Either 3 or 6.}

\item{bw}{Black/white density on the right?}

\item{dens.right.offset}{How much should the right density be moved to the right?}

\item{xlim}{Define limits on x axis}

\item{ylim}{Define limits on y axis}

\item{xextension}{How much reltive space should be on the rights side of the plot? You can adjust this to change the distance of the BF category labels from the plot.}

\item{traj.selection}{Should a fixed set of trajectories be shown ("fixed"), or a selection that reflects the proportional of each stopping category (upper/loewr/n.max hits; use "proportional").}

\item{yaxis.at}{Positions of the ticks on the y-axis. If NA, they are defined automatically.}

\item{yaxis.labels}{Labels of the ticks on the y-axis. If NA, they are defined automatically.}

\item{forH1}{If TRUE, use the BF_10 labels, otherwise the BF_01 labels}

\item{n.max.label.position}{"fixed": Always centered at BF=1. "dynamic": Centered on the the peak of the right density.}
}
\description{
Plot a BPA object
}

