% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{print.BFDAanalysisFixed}
\alias{print.BFDAanalysisFixed}
\title{Print a BFDA analysis}
\usage{
\method{print}{BFDAanalysisFixed}(x, ..., digits = 1)
}
\arguments{
\item{x}{A BFDA-analysis object (which is return from \code{BFDA.analyze})}

\item{...}{(not used)}

\item{digits}{Number of digits in display}
}
\description{
Print a BFDA analysis
}
