% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{print.BPAanalysis}
\alias{print.BPAanalysis}
\title{Print a BPA analysis}
\usage{
\method{print}{BPAanalysis}(x, ..., digits = 1)
}
\arguments{
\item{x}{A BPA-analysis object (which is return from \code{BPA.analysis})}

\item{...}{(not used)}

\item{digits}{Number of digits in display}
}
\description{
Print a BPA analysis
}

