% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtest.R
\name{sample.abtest}
\alias{sample.abtest}
\title{sample.abtest
get two samples with a specified relation of underlying proportions}
\usage{
sample.abtest(n, ES, options.sample = NULL)
}
\arguments{
\item{n}{sample size of the trial per condition}

\item{ES}{effect size (value of odds ratio, log odds ratio, absolute risk, relative risk)}

\item{options.sample}{list defining the type of the input effect size: \code{list(effecttype = ...)} with the options: \code{"OR"} (odds ratio), \code{"RR"} (relative risk), \code{"AR"} (absolute risk), \code{"logOR"} (log odds ratio)}
}
\description{
sample.abtest
get two samples with a specified relation of underlying proportions
}
\examples{
\dontrun{SAMP <- sample.abtest(n=100, ES=2, options.sample = list(effecttype = "OR"))}
}
