% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-Analysis.R
\name{BFDA.analysis}
\alias{BFDA.analysis}
\title{Analyze aBFDA.sim object}
\usage{
BFDA.analysis(BFDA, n.max = NA, boundary = NA, verbose = TRUE,
  alpha = 0.05)
}
\arguments{
\item{BFDA}{The result object from a BFDA.sim function}

\item{n.max}{At which maximum n should the simulations be analyzed?}

\item{boundary}{At which BF boundary should trajectories stop?}

\item{verbose}{Print information about analysis?}

\item{alpha}{For a frequentist analysis in the fixed-n case: Use this alpha level.}
}
\description{
Analyze aBFDA.sim object
}
\examples{
\dontrun{
BFDA.analysis(sim, boundary=6, n.max=80)
}
}

