% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b-evDens.R
\name{evDens}
\alias{evDens}
\title{Plot a BFDA object for fixed n: Two densities for H1 and H0}
\usage{
evDens(BFDA.H1, BFDA.H0, n, boundary = c(1/6, 6), xlim = NA,
  noSplit = FALSE, cex = 1.2, cex.axis = 1, bw = FALSE)
}
\arguments{
\item{BFDA.H1}{A BFDA simulation object for H1, resulting from the \code{BFDA.sim} function}

\item{BFDA.H0}{A BFDA simulation object for H0, resulting from the \code{BFDA.sim} function}

\item{n}{Fixed n at which BF distribution is evaluated}

\item{boundary}{Critical BF boundaries for H0 and H1}

\item{xlim}{limits on xaxis}

\item{noSplit}{If TRUE, the par(mfcol=...) command is skipped}

\item{cex}{Zoom factor for axes labels}

\item{cex.axis}{Zoom factor for tick labels}

\item{bw}{Black/white color scheme? (default:FALSE)}
}
\description{
Plot a BFDA object for fixed n: Two densities for H1 and H0
}
